App structure fix (#9)
