+++ src/components/Footer/constants.ts
@@ -1,0 +1,12 @@
+export const APP_ROUTES = {
+  WELCOME: "/",
+  SIGNIN: "sign_in",
+  SIGNUP: "sign_up",
+  GRAPHIQL: "graphiQL",
+};
+
+export const ProfileLinks = [
+  "https://github.com/ffriday",
+  "https://github.com/krkate",
+  "https://github.com/romankadevich",
+];
