--- src/constants/constants.ts
@@ -1,6 +1,0 @@
-export const APP_ROUTES = {
-  WELCOME: "/",
-  SIGNIN: "sign_in",
-  SIGNUP: "sign_up",
-  GRAPHIQL: "graphiQL",
-};
