+++ src/main.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import { RouterProvider } from "react-router-dom";
-import { AppContextProvider } from "./provisers/LangProvider";
+import { AppContextProvider } from "./providers/LangProvider";
 import { router } from "./router";
-import { AuthProvider } from "./provisers/AuthProviders";
+import { AuthProvider } from "./providers/AuthProviders";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 const queryClient = new QueryClient({
